Fixed Tokenized search for entities, by name, aliases and entity_type, with overall relevance score (float); specs all green; vers. 0.8.2

+++ spec/tools/search_entities_tool_spec.rb
@@ -72,7 +72,7 @@ RSpec.describe SearchEntitiesTool, type: :model do
         )
 
         expect(result[:entity_id]).to be_a(Integer)
-        expect(result[:relevance_score]).to be_a(Integer)
+        expect(result[:relevance_score]).to be_a(Float)
         expect(result[:matched_fields]).to be_an(Array)
       end
